+++ app/helpers/staffs/base_helper.rb
@@ -1,0 +1,2 @@
+module Staffs::BaseHelper
+end
