+++ spec/controllers/staffs/base_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Staffs::BaseController, type: :controller do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
